+++ R/resample.R
@@ -63,17 +63,6 @@ resample = function(task, learner, resampling, store_models = FALSE, store_backe
   assert_flag(store_backends)
   assert_learnable(task, learner)
   set_encapsulation(list(learner), encapsulate)
-  if ("task" %nin% clone) {
-    # ads the test task is replaced anyway with the task from the test rows for learners that use it,
-    # we can replace it here anyway
-    # This ensures that the hashes for the hotstarting are not polluted
-    prev_test_task = task$test_task
-    task$test_task = NULL
-    on.exit({task$test_task = prev_test_task}, add = TRUE)
-  } else {
-    task$test_task = NULL
-  }
-
   if (!resampling$is_instantiated) {
     resampling = resampling$instantiate(task)
   }
@@ -93,7 +82,7 @@ resample = function(task, learner, resampling, store_models = FALSE, store_backe
     hotstart_grid = map_dtr(seq_len(n), function(iteration) {
       if (!is.null(learner$hotstart_stack)) {
         # search for hotstart learner
-        task_hashes = task_hashes(task, resampling)
+        task_hashes = resampling_task_hashes(task, resampling, learner)
         start_learner = get_private(learner$hotstart_stack)$.start_learner(learner$clone(), task_hashes[iteration])
       }
       if (is.null(learner$hotstart_stack) || is.null(start_learner)) {